Debugging print_int

+++ print_int.c
@@ -13,7 +13,7 @@ int print_int(va_list integer, flag *f)
 	long int num;
 	unsigned long int n = 0;
 
-	n = va_arg(integer, long int);
+	n = va_arg(integer, int);
 	num = n;
 	if (f->plus && num >= 0)
 		length += _putchar('+');
